renamed workloads and made code style adjustments
