+++ tests/CMakeLists.txt
@@ -176,7 +176,7 @@ add_fdb_test(TEST_FILES slow/ConfigureTest.txt)
 add_fdb_test(TEST_FILES slow/CycleRollbackPlain.txt)
 add_fdb_test(TEST_FILES slow/DDBalanceAndRemove.txt)
 add_fdb_test(TEST_FILES slow/DDBalanceAndRemoveStatus.txt)
-add_fdb_test(TEST_FILES slow/DDMetrics.txt)
+add_fdb_test(TEST_FILES slow/DDMetricsExclude.txt)
 add_fdb_test(TEST_FILES slow/FastTriggeredWatches.txt)
 add_fdb_test(TEST_FILES slow/LowLatencyWithFailures.txt)
 add_fdb_test(TEST_FILES slow/MoveKeysClean.txt)
